requirements.txt
SimpleWebSocketServerFork==0.1.3
